iommu: msm: ensure lazy mappings are unmapped on detach
Ensure that on detach any lazy mappings for that device are released.

We don't want to leave any outstanding mappings after detach because
the client expects all mappings to have been unmapped, also we don't
want to leave references, in the lazy mapping framework, to devices
which could cease to exists.

CRs-Fixed: 1053605
Change-Id: I9894257c4783073fdba734a175bac486d6ee9347
Signed-off-by: Liam Mark <lmark@codeaurora.org>
Co-authored-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Zile995 <stefan.zivkovic995@gmail.com>