wlan: Delete hash entry on add sta failure
In SAP mode, if add sta fails, association is rejected
and cleanup is done. Hash entry is not removed in the
cleanup path. On receiving auth frame from STA after
this, since hash entry is not deleted, disconnetion will
be triggered and lost_link indication is sent to HDD.
This will result in invalid staidx to be sent in
WLANTL_GetSAPStaRSSi.

Fix is to delete hash entry on add bss failure in
limRejectAssociation.

Change-Id: If2cf85c0f718c915be0b79cbd90c73b6e96a7f6e
CRs-Fixed: 2386866