UPDATE: kernel: Boost whenever a zygote-forked process becomes a top app
Boost the CPU to the max for 1000 ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

"Update this from the old code"

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: kingbri <bdashore3@gmail.com>

signal: fix lmkd check syntax

"The latest sultan update changes this syntax. Sync with this file"

Signed-off-by: kingbri <bdashore3@gmail.com>